LayoutTests/media/modern-media-controls/audio/audio-controls-buttons.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<audio src="../../content/test.mp4" style="width: 600px; height: 50px;" autoplay controls></audio>
<div id="container"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>&lt;audio></code> controls buttons.");

const shadowRoot = document.querySelector("#container");
const media = document.querySelector("audio");
const mediaController = createControls(shadowRoot, media, null);

window.internals.setMockMediaPlaybackTargetPickerEnabled(false);

media.addEventListener("play", () => {
    media.addEventListener("webkitplaybacktargetavailabilitychanged", playbackTargetAvailabilityChanged);

    debug("Making AirPlay routes available.");
    window.internals.setMockMediaPlaybackTargetPickerEnabled(true);
});

function playbackTargetAvailabilityChanged(event)
{
    if (event.availability != "available")
        return;

    debug("AirPlay routes became available.");

    scheduler.frameDidFire = function()
    {
        shouldBe("mediaController.controls.controlsBar.children[2].children.length", "2");
        shouldBe("mediaController.controls.controlsBar.children[2].children[0]", "mediaController.controls.muteButton");
        shouldBe("mediaController.controls.controlsBar.children[2].children[1]", "mediaController.controls.airplayButton");

        debug("");
        shadowRoot.remove();
        media.remove();
        finishJSTest();
    };
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
